Move `ty::{Placeholder, PlaceholderConst, PlaceholderRegion, PlaceholderType}` to their own little module (cute)
